Now calculating ngrams with C++ instead of awk.
The C++ ngram counting script (located in cngrams/) is precompiled and is used just like the awk script was used: run a shell command by passing arguments in a subprocess. Also just like the awk script, it's generalized for any size N.

In a speed comparison, the C++ script is 30-40% faster than awk.